pyproject.toml
[project]
name = "perplexity-webui-scraper"
version = "0.2.0"
description = "Perplexity WebUI Scraper"
readme = "README.md"
authors = [
    { name = "henrique-coder", email = "henriquemoreira10fk@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "curl-cffi>=0.13.0",
    "httpx[http2]>=0.28.1",
    "orjson>=3.11.4",
    "pydantic>=2.12.4",
]

[dependency-groups]
dev = [
    "camoufox>=0.4.11",
    "fastapi>=0.121.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "uvicorn>=0.38.0",
]
lint = [
    "ruff>=0.14.6",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"
